added IPv6 support for single /128 addresses of routers (based on IPv4 address)
git-svn-id: https://dev.arachne.cz/repos/prometheus/trunk@246 251d49ef-1d17-4917-a970-b30cf55b089b

+++ ipstruct.h
@@ -42,7 +42,7 @@ struct Keyword
 {
  char *key;
  
- int asymetry_ratio;        /* ratio for ADSL-like upload */
+ float asymetry_ratio;      /* ratio for ADSL-like upload */
  int asymetry_fixed;        /* fixed treshold for ADSL-like upload */
  int data_limit;            /* hard shaping: apply magic_treshold if max*data_limit MB exceeded */
  int data_prio;             /* soft shaping (qos): reduce HTB prio if max*data_prio MB exceeded */
